Split progress snapshot calculations
